Extract the printing types into separate modules
I had put all the printing types and logic into a single file (`printing.rs`). This pulls things out into three separating files (`print.rs`, `print_char.rs`, and `print_string.rs`).
