+++ packages/push/src/instruction/int/mod.rs
@@ -13,7 +13,7 @@ use super::{
         dup::Dup, flush::Flush, is_empty::IsEmpty, pop::Pop, push_value::PushValue,
         stack_depth::StackDepth, swap::Swap,
     },
-    printing::Print,
+    printing::print::Print,
 };
 use crate::{
     error::{Error, InstructionResult},
