+++ packages/push/src/instruction/printing/mod.rs
@@ -1,0 +1,3 @@
+pub mod print;
+pub mod print_char;
+pub mod print_string;
